requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
duckdb==1.5.0.dev94
entrypoints==0.4
Faker==37.12.0
favicon==0.7.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.3
groq==0.33.0
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.1
gspread-pandas==3.3.0
h11==0.16.0
hf-xet==1.2.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
mysql-connector-python==9.5.0
narwhals==2.10.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pillow==11.3.0
plotly==6.3.1
prometheus_client==0.23.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pymdown-extensions==10.16.1
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sql_metadata==2.18.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
st-annotated-text==4.0.2
st-gsheets-connection==0.1.0
st-theme==1.2.3
streamlit==1.50.0
streamlit-autorefresh==1.0.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
